구매목록
2023-10-17
Error : Int 초과 -> Long 해결

+++ FC_Online/app/src/main/java/com/example/fc_online/data/TradeType.kt
@@ -12,5 +12,5 @@ data class TradeType( // 거래기록
     @SerializedName("grade") // 선수 강화등급
     val grade: Int = 0,
     @SerializedName("value") // 선수 가치(BP)
-    val value: Int = 0
+    val value: Long = 0
 )
